Fixes filename and folder problems
Files were not getting saved in root directory of bucket.
Added in today's date as a folder name prepending file name.
Uses crypto.randomBytes to create a random string for title on AWS.
Promisify both s3.upload and crypto.randomBytes.
Both implement promises that are resolved or rejected on if they work.